re add the client_mock.go
