+++ ika/client_test.go
@@ -5,32 +5,11 @@ import (
 	"os"
 	"testing"
 
-	"github.com/block-vision/sui-go-sdk/models"
 	"github.com/block-vision/sui-go-sdk/signer"
 	"github.com/block-vision/sui-go-sdk/sui"
-	tmtypes "github.com/cometbft/cometbft/types"
 	"github.com/joho/godotenv"
-	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/assert"
 )
-
-// MockClient is a mock implementation of the Client interface for testing.
-type MockClient struct{}
-
-// NewMockClient creates a new mock Client instance.
-func NewMockClient() *MockClient {
-	return &MockClient{}
-}
-
-// UpdateLC is a mock implementation used in tests.
-func (p *MockClient) UpdateLC(_ context.Context, _ *tmtypes.LightBlock, _ zerolog.Logger) (models.SuiTransactionBlockResponse, error) {
-	return models.SuiTransactionBlockResponse{}, nil
-}
-
-// ApproveAndSign is a mock implementation that returns a predefined signature.
-func (p *MockClient) ApproveAndSign(_ context.Context, _, _ string, _ [][]byte) ([][]byte, error) {
-	return [][]byte{{0x01, 0x02, 0x03}}, nil
-}
 
 func TestClient_ApproveAndSign(t *testing.T) {
 	t.Skip("Test to be run locally for debugging purposes only")
